chore(weave): merge link_prompt_to_registry test files into one
Consolidate transport and client tests into a single file at
tests/trace/test_link_prompt_to_registry.py. All 15 test cases preserved.

Made-with: Cursor
